backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
elevenlabs
httpx
python-dotenv
google-generativeai
aiosqlite
elevenlabs
